refactor: update route directory
