+++ src/ServiceProviders/RestAPI.php
@@ -27,7 +27,7 @@ use Give\Donations\Endpoints\SwitchDonationView;
 use Give\Donors\Endpoints\DeleteDonor;
 use Give\Donors\Endpoints\ListDonors;
 use Give\Donors\Endpoints\SwitchDonorView;
-use Give\Promotions\InPluginUpsells\Endpoints\ProductRecommendations;
+use Give\Promotions\InPluginUpsells\ProductRecommendationsRoute;
 use Give\Subscriptions\Endpoints\ListSubscriptions;
 use Give\Subscriptions\Endpoints\SubscriptionActions;
 use Give\Subscriptions\Endpoints\SwitchSubscriptionView;
@@ -65,7 +65,7 @@ class RestAPI implements ServiceProvider
         FormActions::class,
         GetMigrations::class,
         RunMigration::class,
-        ProductRecommendations::class,
+        ProductRecommendationsRoute::class,
     ];
 
     /**
